# internal/models/user.go
/*
Copyright (C) 2025 Your Company
All Rights Reserved.
*/
// internal/models/user.go
package models

import "time"

// User represents a user in the system.
// @swagger:model User
type User struct {
	ID           string    `gorm:"type:uuid;primaryKey" json:"id"`
	Username     string    `gorm:"type:varchar(100);unique;not null" json:"username"`
	Email        string    `gorm:"type:varchar(255);unique;not null" json:"email"`
	PasswordHash string    `gorm:"type:text;not null" json:"-"`
	Preferences  string    `gorm:"type:jsonb" json:"preferences"`
	CreatedAt    time.Time `gorm:"autoCreateTime" json:"created_at"`
	UpdatedAt    time.Time `gorm:"autoUpdateTime" json:"updated_at"`
}